SplashScreen in the final stage of Development
Changed the way of showing the background. Now using a dedicated xml resource file (background_splash.xml) as the Theme;
Developed the "Downloading" label (include the creation of a new String);
Developed the 3 dots animation using a GitHub library;
Created a simple animation to make a "slide to left" transition from the SplashScreen Activity to the Main Activity.

Also:
Created the new Icon;
Changed the Articles Database to include the "Read" boolean - to allow the user mark the article as read;
Changed the ArticlesAdapter to reflect the "Read" in the Database to the layout;
Fixed the Sorting of website for Ignore Case (because of websites like iMore).