- Changed the Layout for an Adaptive Layout -
Created the xlarge layout files for the screens that need changes when using in a greater device (tablet) - Read Screen;
Created the land layout files for the screens that need changes to fit better when using the device in landscape mode - Main Screen;
Changed the Adapter to perform the right way (with holder);
Implemented the onConfigurationChanged in the Activities to handle the screen orientation changes;
Changed the Android Manifest to allow the configuration changed listener;
Some improvements in layouts.

Also:
Removed the Image Caching indicators
Fixed the spaces on the end of some Strings

To future:
Develop the Tablet layout in a way to List and Read in the same Screen. In the way that I don't own a tablet for tests, it's risky to do by now.